+++ src/hal/project_specific_crash_handler.cpp
@@ -4,6 +4,14 @@
 
 #include <cstring>
 
+#if defined(_WIN32)
+#define WIN32_LEAN_AND_MEAN
+#ifndef NOMINMAX
+#define NOMINMAX
+#endif
+#include <windows.h>
+#endif
+
 using namespace das;
 
 // Separate function because C++ objects (std::string) cannot coexist with
